Use itemized data for storing selected customer price.
Itemized data allows a customer to add more than one variant of a customer
priced product at the same time. Additionally, by not storing it in the session
it will automatically be cleared when the product is removed from the cart.

This also lets customer pricing be used for when a shop needs only very basic
variant support.